backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Excel
openpyxl>=3.1.2

# HTTP client (for WhatsApp etc.)
httpx>=0.26.0
twilio>=8.0.0

# Email (password reset; used when EMAIL_BACKEND=resend)
resend>=2.0.0
